test: merge ang. dist. tests into one file
